Allows only owners of the image to edit/delete them

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,15 +92,25 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
+    // This method also validates that the user trying to edit the image is the owner. If the User isn't the owner of the image, he/she is prevented and
+    // an appropriate error message is displayed
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
+        User imageUser = (User)session.getAttribute("loggeduser");
+        String error = "Only the owner of the image can edit the image";
         model.addAttribute("image", image);
+        if(image.getUser().getId() != imageUser.getId()){
+
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+        String tags = convertTagsToString(image.getTags());
         model.addAttribute("tags", tags);
+
         return "images/edit";
     }
+
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
     //The method receives the imageFile, imageId, updated image, along with the Http Session
@@ -140,8 +150,20 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+    // This method validates if the user trying to delete the image is the owner of the image; Only if the user is the owner of the image
+    // can they proceed to delete, else an error message is displayed.
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        User imageUser = (User)session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        String error = "Only the owner of the image can delete the image";
+
+        if (image.getUser().getId() != imageUser.getId()) {
+            model.addAttribute("deleteError", error);
+            model.addAttribute("image", image);
+            return "images/image";
+        }
+        //imageService.deleteImage(imageId,imageUser);
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
